requirements.txt
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
cryptography==36.0.2
idna==2.7
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.4.2
python-dotenv==0.20.0
requests==2.20.0
toml==0.10.2
urllib3==1.24.3
